pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "superset-showtime"
version = "0.1.0"
description = "🎪 Apache Superset ephemeral environment management with circus tent emoji state tracking"
readme = "README.md"
license = "Apache-2.0"
requires-python = ">=3.8"
authors = [
    { name = "Maxime Beauchemin", email = "maximebeauchemin@gmail.com" },
]
keywords = [
    "ephemeral",
    "environments",
    "devops",
    "aws",
    "github",
    "superset",
    "circus",
    "showtime"
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Operating System :: OS Independent",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Build Tools",
    "Topic :: System :: Systems Administration",
]
dependencies = [
    "typer>=0.9.0",
    "rich>=10.0.0",
    "pydantic>=2.0.0",
    "gitpython>=3.1.0",
    "httpx>=0.24.0",
    "pyyaml>=6.0",
    "python-dateutil>=2.8.0",
]

[project.optional-dependencies]
aws = [
    "boto3>=1.26.0",
    "botocore>=1.29.0",
]
gcp = [
    "google-cloud-compute>=1.11.0",
    "google-cloud-storage>=2.8.0",
]
azure = [
    "azure-mgmt-containerinstance>=10.0.0",
    "azure-storage-blob>=12.0.0",
]
k8s = [
    "kubernetes>=26.0.0",
]
all = [
    "superset-showtime[aws,gcp,azure,k8s]",
]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "ruff>=0.1.9",
    "mypy>=1.0.0",
    "pre-commit>=3.0.0",
]

[project.urls]
Homepage = "https://github.com/apache/superset-showtime"
Documentation = "https://superset-showtime.readthedocs.io/"
Repository = "https://github.com/apache/superset-showtime.git"
"Bug Tracker" = "https://github.com/apache/superset-showtime/issues"
Changelog = "https://github.com/apache/superset-showtime/blob/main/CHANGELOG.md"

[project.scripts]
superset-showtime = "showtime.cli:main"
showtime = "showtime.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["showtime"]

[tool.ruff]
target-version = "py38"
line-length = 100

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
]
ignore = [
    "E501", # line too long
    "B008", # do not perform function calls in argument defaults
    "C901", # too complex
]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true

[tool.pytest.ini_options]
minversion = "7.0"
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
addopts = "--cov=showtime --cov-report=term-missing --cov-report=xml"
